Mute warning

+++ testing/span.cpp
@@ -8,7 +8,7 @@ TEST_CASE("Check uniformity")
     dbs::Span span = {{0, 0}, 1, {0, 1}};
     for(int i = 0; i <= B; i++)
     {
-        span.bulge = i;
+        span.bulge = (float)i;
         float min, max;
         dbs::Complex p, prev;
         for(int j = 0; j <= N; j++, prev = p)
